Rudimentary Multiplayer mechanism
The world scene has been given signals to emit vehicle controls, as well
as the input scanning to emit those signals. The tank has lost it's
input scanning and instead has functions meant to be used with the world
scene's signals.

A multiplayer test scene has been set up -- mp_test.tscn -- to test the
very basics of Godot multiplayer. It syncs a couple of objects, and only
provides a very basic connection UI. But look! It does the thing!